requirements.txt
python-dotenv
tweepy
beautifulsoup4
schedule
requests
